refactor: convert client and server to libraries
