+++ Client/client.go
@@ -1,4 +1,4 @@
-package main
+package client
 
 import (
 	"E2E2/Cipher"
@@ -166,24 +166,4 @@ func (c *Client) Send(message string) (string, error) {
 	}
 
 	return decrypted, nil
-}
-
-func main() {
-	client, err := NewClient("http://localhost:8080")
-	if err != nil {
-		log.Fatalf("Failed to create client: %v", err)
-	}
-
-	if !client.LoadSession() {
-		if err := client.ExchangeKeys(); err != nil {
-			log.Fatalf("Failed to exchange keys: %v", err)
-		}
-	}
-
-	response, err := client.Send("ping")
-	if err != nil {
-		log.Fatalf("Failed to send message: %v", err)
-	}
-
-	fmt.Printf("Server response: %s\n", response)
 }
