--- Client/client_session.json
@@ -1,1 +1,0 @@
-{"APIKey":"f009e8bb-8d1d-4efa-9d97-466adefd0f87","K3":"96c686cc57996e0e56d510a92e6e40171dea1b2fe2339842bfd0bb423010eeeb","SessionID":"27e0b5d1-face-4d1d-9d06-a0594001f981"}
