+++ Server/server.go
@@ -1,11 +1,10 @@
-package main
+package server
 
 import (
 	"E2E2/Cipher"
 	"E2E2/Storage"
 	"encoding/hex"
 	"fmt"
-	"log"
 	"net/http"
 	"sync"
 	"time"
@@ -184,24 +183,4 @@ func (s *Server) HandleTunnel(callback func(string) string) gin.HandlerFunc {
 
 		c.JSON(http.StatusOK, gin.H{"Data": encrypted})
 	}
-}
-
-func main() {
-	server, err := NewServer()
-	if err != nil {
-		log.Fatalf("Failed to create server: %v", err)
-	}
-
-	router := gin.Default()
-	router.POST("/exchange-keys", server.HandleKeyExchange)
-
-	router.POST("/tunnel", server.HandleTunnel(func(response string) string {
-		if response == "ping" {
-			return "pong"
-		}
-		return "unknown message"
-	}))
-
-	log.Println("Server starting on :8080")
-	log.Fatal(router.Run(":8080"))
 }
