--- Server/server_sessions.json
@@ -1,5 +1,0 @@
-{
-    "3bc14209-de45-48b8-b7d5-2ae6a813cb60": "c35c739b1265f956a18dd2388e5ca2460453070af256414c02a1c23022680a52",
-    "bf7169b8-77b4-4a4b-9ec9-480272818bc1": "9b3a8875fd4245b2b411d21cfc85f90b9565d38f5c383daef0b07783bf9cba42",
-    "f009e8bb-8d1d-4efa-9d97-466adefd0f87": "837a047b8f0b98f3f30bc276dd30db791780edf865f188955d7b7c5446f6f4d2"
-}
